Split RankOneTensor class
