+++ lib/src/Uncertainty/Algorithm/MetaModel/TensorApproximation/CMakeLists.txt
@@ -3,6 +3,7 @@
 # Register current directory files
 ot_add_current_dir_to_include_dirs ()
 
+ot_add_source_file (RankOneTensor.cxx)
 ot_add_source_file (CanonicalTensor.cxx)
 ot_add_source_file (RankOneTensorEvaluation.cxx)
 ot_add_source_file (TensorApproximationResult.cxx)
@@ -10,6 +11,7 @@ ot_add_source_file (TensorApproximationAlgorithm.cxx)
 
 ot_install_header_file (OTTensorApproximation.hxx)
 
+ot_install_header_file (RankOneTensor.hxx)
 ot_install_header_file (CanonicalTensor.hxx)
 ot_install_header_file (RankOneTensorEvaluation.hxx)
 ot_install_header_file (TensorApproximationResult.hxx)
