+++ lib/src/Uncertainty/Algorithm/MetaModel/TensorApproximation/openturns/RankOneTensor.hxx
@@ -1,6 +1,6 @@
 //                                               -*- C++ -*-
 /**
- *  @brief Canonical tensor format
+ *  @brief Rank-1 tensor
  *
  *  Copyright 2005-2016 Airbus-EDF-IMACS-Phimeca
  *
@@ -18,44 +18,27 @@
  *  along with this library.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
-#ifndef OPENTURNS_CANONICALTENSOR_HXX
-#define OPENTURNS_CANONICALTENSOR_HXX
+#ifndef OPENTURNS_RANKONETENSOR_HXX
+#define OPENTURNS_RANKONETENSOR_HXX
 
 #include "openturns/OrthogonalUniVariateFunctionFamily.hxx"
-#include "openturns/RankOneTensor.hxx"
+#include "openturns/Basis.hxx"
 
 BEGIN_NAMESPACE_OPENTURNS
 
-
-/**
- * @class CanonicalTensor
- * Canonical tensor format
- */
-class OT_API CanonicalTensor
-  : public PersistentObject
+class OT_API RankOneTensor : public PersistentObject
 {
-
   CLASSNAME;
 public:
-  typedef Collection<OrthogonalUniVariateFunctionFamily>           FunctionFamilyCollection;
+
+  typedef Collection<OrthogonalUniVariateFunctionFamily> FunctionFamilyCollection;
   typedef PersistentCollection<OrthogonalUniVariateFunctionFamily> FunctionFamilyPersistentCollection;
 
-  /** Default constructor */
-  CanonicalTensor();
-
-  /** Standard constructor */
-  CanonicalTensor(const FunctionFamilyCollection & functionFamilies,
-                  const Indices & nk,
-                  const UnsignedInteger rank = 1);
-
-  void setRank(const UnsignedInteger rank);
-  UnsignedInteger getRank() const;
+  explicit RankOneTensor(const FunctionFamilyCollection & functionFamilies = FunctionFamilyCollection(0),
+                         const Indices & nk = Indices());
 
   /** Virtual constructor */
-  virtual CanonicalTensor * clone() const;
-
-  /** String converter */
-  virtual String __repr__() const;
+  virtual RankOneTensor * clone() const;
 
   /** Method save() stores the object through the StorageManager */
   virtual void save(Advocate & adv) const;
@@ -66,13 +49,22 @@ public:
 protected:
   friend class TensorApproximationAlgorithm;
   friend class TensorApproximationResult;
+  friend class RankOneTensorEvaluation;
+
+  // constant factor
+  NumericalScalar radius_;
+
+  // subbasis sizes
+  Indices nk_;
+
+  // subbasis coefficients
+  PersistentCollection<NumericalPoint> coefficients_;
 
   FunctionFamilyPersistentCollection functionFamilies_;
+  PersistentCollection<Basis> basis_;
+};
 
-  Collection<RankOneTensor> rank1tensors_;
-
-}; // class CanonicalTensor
 
 END_NAMESPACE_OPENTURNS
 
-#endif /* OPENTURNS_CANONICALTENSOR_HXX */
+#endif /* OPENTURNS_RANKONETENSOR_HXX */
